bone_configurations_drag_rotate_detent_bone now tests what it should test and set_animation_data applies SHOW/HIDE animations properly. And location keyframes on a bone's group is fixed. Not only does it pass, but *arm and *bone give the same messages.
In addition, a lesson learned from making this "easy thin wrapper": Programmatically creating XPlane2Blender + Blender data is proving to be full of lots of little issues thanks to Blender's "Needle In A Pile Of Pasta And Yarn - You Can't Get There From Here" data model. For future programmers who are extending this "thin easy wrapper" or are attempting to use blenderObject and blenderBone information for validation messages: You're not crazy, this is just how it is. Start by reasoning what you need, then when it can be available to you! Then, prepare to do 2x the work if your input/output accepts an Object or a Bone. You're essentatially developing the same feature twice.