Fix bug #351 - bone animation parenting data block is not WYSIWYG
The old exporter would find the pre-animation matrix for DBs by walking
forward from the parent and simulating the parent-child process.  This
simulation was wrong for bones and generally error-prone since it
required us to duplicate a big chunk of opaque Blender code.

The new method walks backward from the datablock pose; since data block
animations are last in Blender's transform algo, this is actually a lot
less work, and more reliable.