More commitment to mtpoly's mode value, clean up of comments, logger text, and importantly a semantic change in _ModMember - it just means pressed on or off. We decide later what that means.
We also have, for the first time a real unit test for this feature. It is hardly completed, and some parts won't be completed till much later, and it is much more expanded than what I thought it would be, but, it is very nice. It tests if the properties are al what they should be and only what should have changed did.