Ensures that the DEFAULT_MATERIAL_NAME is properly appended without a context issue and that autogenerated materials also get the specular fix.
btw, if you're going through the logs in a bisect going "how was he testing this?  How was this working before that was?" This was developed very very quickly. I'm trying to commit the bits and pieces that are special and working seperately to keep track of this very long complex evolving conversion.